@@ -1,155 +1,29 @@
 {
-  "name": "finai-stock-analysis-client",
-  "version": "1.0.0",
-  "repository": {
-    "type": "git",
-    "url": "https://github.com/Preston-C89-Chen/app.stock_analysis"
-  },
-  "author": "Preston Chen <pchen415@gmail.com>",
-  "license": "MIT",
+  "name": "mosaic-react",
+  "version": "0.1.0",
   "scripts": {
-    "dev-client": "cross-env NODE_ENV=development webpack-dev-server --hot --config webpack.config.dev.js"
+    "dev": "vite",
+    "build": "vite build",
+    "preview": "vite preview"
   },
   "dependencies": {
-    "@babel/cli": "^7.2.3",
-    "@babel/core": "^7.2.2",
-    "@babel/plugin-proposal-class-properties": "^7.3.0",
-    "@babel/plugin-syntax-dynamic-import": "^7.2.0",
-    "@babel/plugin-transform-runtime": "^7.2.0",
-    "@babel/preset-env": "^7.3.1",
-    "@babel/preset-react": "^7.0.0",
-    "@babel/preset-typescript": "^7.1.0",
     "@radix-ui/colors": "^3.0.0",
-    "@radix-ui/react-icons": "^1.3.0",
-    "@radix-ui/react-popover": "1.0.7",
     "@radix-ui/themes": "^2.0.0",
-    "@types/i18next": "^12.1.0",
-    "@types/i18next-browser-languagedetector": "^2.0.1",
-    "@types/i18next-xhr-backend": "^1.4.1",
-    "@types/materialize-css": "^1.0.6",
-    "@types/prismjs": "^1.9.0",
-    "@types/react": "^16.8.2",
-    "@types/react-content-loader": "^3.1.4",
-    "@types/react-dom": "^16.8.0",
-    "@types/react-redux": "^7.0.1",
-    "@types/react-router": "^4.4.3",
-    "@types/react-router-dom": "^4.3.1",
-    "@types/redux-immutable": "^4.0.1",
-    "@types/redux-logger": "^3.0.6",
-    "@types/uuid": "^3.4.4",
-    "@types/webpack-env": "^1.13.7",
-    "add-asset-html-webpack-plugin": "^3.1.3",
-    "awesome-typescript-loader": "^5.2.1",
-    "axios": "^0.18.0",
-    "babel-loader": "^8.0.5",
-    "babel-plugin-prismjs": "^1.0.2",
-    "colors": "^1.3.3",
-    "compression": "^1.7.3",
-    "connect-redis": "^3.4.0",
-    "connected-react-router": "^6.2.2",
-    "cookie-parser": "^1.4.3",
-    "copy-webpack-plugin": "^4.6.0",
-    "cross-env": "^5.2.0",
-    "css-loader": "^2.1.0",
-    "cssnano": "^4.1.8",
-    "dotenv-webpack": "^1.7.0",
-    "duplicate-package-checker-webpack-plugin": "^3.0.0",
-    "file-loader": "^3.0.1",
-    "fork-ts-checker-webpack-plugin": "^0.5.2",
-    "helmet": "^3.15.0",
-    "history": "^4.7.2",
-    "html-webpack-plugin": "^3.2.0",
-    "http-status": "^1.3.1",
-    "i18next": "^14.1.1",
-    "i18next-browser-languagedetector": "^2.2.4",
-    "i18next-xhr-backend": "^1.5.1",
-    "image-webpack-loader": "^4.6.0",
-    "immutable": "4.0.0-rc.12",
-    "mini-css-extract-plugin": "^0.5.0",
-    "node-sass": "^6.0.0",
-    "offline-plugin": "^5.0.6",
-    "pg": "^7.8.0",
-    "postcss": "^7.0.14",
-    "postcss-import": "^12.0.1",
-    "postcss-loader": "^3.0.0",
-    "postcss-preset-env": "^6.5.0",
-    "prismjs": "^1.15.0",
-    "progress-bar-webpack-plugin": "^1.12.1",
-    "react": "^16.8.1",
-    "react-content-loader": "^4.0.1",
-    "react-dom": "^16.8.1",
-    "react-hot-loader": "^4.6.5",
-    "react-i18next": "^10.0.1",
-    "react-redux": "^6.0.0",
-    "react-router": "^4.3.1",
-    "react-router-dom": "^6.16.0",
-    "redux": "^4.0.1",
-    "redux-immutable": "^4.0.0",
-    "redux-logger": "^3.0.6",
-    "redux-saga": "^1.0.1",
-    "sass-loader": "^7.1.0",
-    "style-loader": "^0.23.1",
-    "trash-cli": "^1.4.0",
-    "ts-loader": "^5.3.3",
-    "typescript": "^5.2.2",
-    "url-loader": "^1.1.2",
-    "uuid": "^3.3.2",
-    "webpack": "^4.29.3",
-    "webpack-bundle-analyzer": "^3.0.3",
-    "webpack-cli": "^3.2.3",
-    "webpack-merge": "^4.2.1",
-    "webpack-pwa-manifest": "^4.0.0"
+    "@tailwindcss/forms": "^0.5.3",
+    "chart.js": "^4.3.0",
+    "chartjs-adapter-moment": "^1.0.1",
+    "moment": "^2.29.4",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "react-flatpickr": "^3.10.13",
+    "react-router-dom": "^6.11.2",
+    "react-transition-group": "^4.4.5"
   },
   "devDependencies": {
-    "@babel/node": "^7.2.2",
-    "@types/jest": "^24.0.0",
-    "@types/react-test-renderer": "^16.8.0",
-    "babel-core": "7.0.0-bridge.0",
-    "case-sensitive-paths-webpack-plugin": "^2.2.0",
-    "commitlint": "^7.5.0",
-    "commitlint-config-armour": "^1.2.1",
-    "coveralls": "^3.0.2",
-    "dotenv": "^6.2.0",
-    "eslint": "^5.13.0",
-    "eslint-config-airbnb": "^17.1.0",
-    "eslint-import-resolver-webpack": "^0.11.0",
-    "eslint-plugin-import": "^2.16.0",
-    "eslint-plugin-jsx-a11y": "^6.2.1",
-    "eslint-plugin-react": "^7.12.4",
-    "husky": "1.3.1",
-    "identity-obj-proxy": "^3.0.0",
-    "jest": "^24.1.0",
-    "react-test-renderer": "^16.8.1",
-    "stylelint": "^9.10.1",
-    "stylelint-config-standard": "^18.2.0",
-    "ts-jest": "^23.10.5",
-    "tslint": "^5.12.1",
-    "tslint-react": "^3.6.0",
-    "webpack-dev-server": "^3.1.14"
-  },
-  "resolutions": {
-    "**/event-stream": "^4.0.1"
-  },
-  "browserslist": [
-    "> 1%",
-    "last 2 versions"
-  ],
-  "jest": {
-    "preset": "ts-jest",
-    "moduleNameMapper": {
-      "\\.(jpg|jpeg|png|gif|eot|otf|webp|svg|ttf|woff|woff2|mp4|webm|wav|mp3|m4a|aac|oga)$": "<rootDir>/__mocks__/fileMock.tsx",
-      "\\.(css|less|scss|sass)$": "identity-obj-proxy"
-    },
-    "moduleDirectories": [
-      "node_modules",
-      "frontend/src"
-    ],
-    "moduleFileExtensions": [
-      "js",
-      "jsx",
-      "json",
-      "ts",
-      "tsx"
-    ]
+    "@vitejs/plugin-react": "^4.1.0",
+    "autoprefixer": "^10.4.14",
+    "postcss": "^8.4.24",
+    "tailwindcss": "^3.3.2",
+    "vite": "^4.3.9"
   }
 }
