@@ -1,6 +1,7 @@
-module.exports = {
-  plugins: [
-    require('tailwindcss'),
-    require('autoprefixer'),
-  ],
-};
+import tailwind from 'tailwindcss'
+import autoprefixer from 'autoprefixer'
+import tailwindConfig from './src/css/tailwind.config.js'
+
+export default {
+  plugins: [tailwind(tailwindConfig), autoprefixer],
+}
